feat: added accelerated media switch
